LayoutTests/imported/w3c/web-platform-tests/fetch/content-type/script.window-expected.txt


PASS Loading JSON… 
PASS separate text/javascript;charset=windows-1252 
PASS separate text/javascript;";charset=windows-1252 
FAIL separate text/javascript assert_unreached: Reached unreachable code
PASS separate "text/javascript" 
PASS separate text/ javascript 
PASS separate text /javascript 
PASS separate x/x text/javascript 
FAIL combined x/x text/javascript assert_unreached: Reached unreachable code
PASS separate x/x;charset=windows-1252 text/javascript 
FAIL combined x/x;charset=windows-1252 text/javascript assert_unreached: Reached unreachable code
PASS separate text/javascript x/x 
FAIL combined text/javascript x/x assert_unreached: Reached unreachable code
FAIL separate text/javascript; charset=windows-1252 text/javascript assert_equals: expected "â‚¬" but got "€"
PASS combined text/javascript; charset=windows-1252 text/javascript 
FAIL separate text/javascript;" x/x assert_unreached: Reached unreachable code
PASS combined text/javascript;" x/x 
FAIL separate text/javascript  assert_unreached: Reached unreachable code
PASS combined text/javascript  
FAIL separate text/javascript error assert_unreached: Reached unreachable code
PASS combined text/javascript error 
PASS separate text/javascript;charset=windows-1252 x/x text/javascript 
FAIL combined text/javascript;charset=windows-1252 x/x text/javascript assert_equals: expected "€" but got "â‚¬"
FAIL separate text/javascript;charset=windows-1252 error text/javascript assert_equals: expected "â‚¬" but got "€"
PASS combined text/javascript;charset=windows-1252 error text/javascript 
FAIL separate text/javascript;charset=windows-1252  text/javascript assert_equals: expected "â‚¬" but got "€"
PASS combined text/javascript;charset=windows-1252  text/javascript 
FAIL separate text/javascript;charset=windows-1252;" \" x/x assert_unreached: Reached unreachable code
PASS combined text/javascript;charset=windows-1252;" \" x/x 
PASS separate x/x;" x/y;\" text/javascript;charset=windows-1252;" text/javascript 
FAIL combined x/x;" x/y;\" text/javascript;charset=windows-1252;" text/javascript assert_unreached: Reached unreachable code
